Removing expiresOn from RefreshToken classes. Updating tests.
